feat: update document controller tests

+++ app/Http/Controllers/Documents/DocumentController.php
@@ -15,6 +15,7 @@ use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 use Inertia\Response;
+use Symfony\Component\HttpFoundation\Response as HttpStatus;
 
 final class DocumentController
 {
@@ -77,6 +78,8 @@ final class DocumentController
      */
     public function show(Document $document): Response
     {
+        abort_if($document->user_id !== $this->verifiedUser()->id, HttpStatus::HTTP_NOT_FOUND);
+
         $documentData = $document
             ->load([
                 'analysis',
@@ -94,6 +97,8 @@ final class DocumentController
      */
     public function edit(Document $document): Response
     {
+        abort_if($document->user_id !== $this->verifiedUser()->id, HttpStatus::HTTP_NOT_FOUND);
+
         return Inertia::render('documents/edit', [
             'document' => DocumentSummaryData::from($document->load('analysis')),
         ]);
@@ -104,6 +109,8 @@ final class DocumentController
      */
     public function update(UpdateDocumentRequest $request, Document $document): RedirectResponse
     {
+        abort_if($document->user_id !== $this->verifiedUser()->id, HttpStatus::HTTP_NOT_FOUND);
+
         $document->update([
             'name' => $request->input('name'),
             'description' => $request->input('description'),
@@ -119,6 +126,8 @@ final class DocumentController
      */
     public function destroy(Document $document): RedirectResponse
     {
+        abort_if($document->user_id !== $this->verifiedUser()->id, HttpStatus::HTTP_NOT_FOUND);
+
         Storage::delete($document->path);
 
         if ($document->thumbnail !== null) {
